feat: load TigerGraph connection config from YAML file
