+++ tests/integration/core/config/tigergraph_connection.yaml
@@ -1,0 +1,3 @@
+host: "http://34.31.117.27"
+username: "tigergraph"
+password: "tigergraph"
