+++ tests/integration/core/tigergraph_api/tigergraph_api_test.py
@@ -1,6 +1,8 @@
 import pytest
 from requests.exceptions import ConnectionError
 from pydantic import HttpUrl
+from pathlib import Path
+import yaml
 
 from tigergraphx.core import Graph
 from tigergraphx.core.tigergraph_api import TigerGraphAPI, TigerGraphAPIError
@@ -37,10 +39,16 @@ class TestTigerGraphAPI:
                 },
             },
         }
+        # Load config from YAML
+        config_path = Path(__file__).parent.parent / "config" / "tigergraph_connection.yaml"
+        with open(config_path, "r") as f:
+            config_dict = yaml.safe_load(f)
+
+        # Parse with TigerGraphConnectionConfig
         self.tigergraph_connection_config = TigerGraphConnectionConfig(
-            host=HttpUrl("http://34.31.117.27"),
-            username="tigergraph",
-            password="tigergraph",
+            host=HttpUrl(config_dict["host"]),
+            username=config_dict["username"],
+            password=config_dict["password"],
         )
         self.G = Graph(
             graph_schema=graph_schema,
